requirements.txt
streamlit
langchain
python-dotenv
faiss-cpu
pyairtable
sentence-transformers
instructorembedding
pinecone-client
